Display howto confirm dialog only when form is dirty
When leaving the howto edit page a confirmation dialog is shown to warn
the user that he/she might loose the changes. This commit changes this
logic so that the dialog is only shown when there are actual changes
that would be lost.
The dirty state of the form is used for this. A big part of the added
logic is for correctly marking a field (and thus the form) as dirty.