feat: adds redirect for 404 how tos
There are many reasons a how to may not be available
on the server. Such as:

* Deleted :fire:
* Renamed :pencil2:

In an effort to make the experience less of a dead
end for the visitor we redirect them to the search
page.

This current implementation takes place on the client
side after attempting to load a how-to entry from
the backend data store.

Reference:
https://github.com/ONEARMY/community-platform/issues/1168