feat: load log level from the environment
The logging library will only output logs
which meet or exceed the defined log level.

For example if the value is set to `debug`
then `debug,info,warn,error` will be emitted.
This is likely the preferred setup for local
development.

In production however we probably want to default
to something like `warn` which will ensure that
only those logs which represent users reaching
a problematic state in the application.