fix: hide user profile section
The User Profile section has quite a lot of information
which is tied to specifics of the original Precious Plastic
deployment. The change here is a quick fix rather than
a long term solution.

Ideally the concept of user focus' can be removed from the
code and made into an extensible object.