chore: reduce react icon bundle size
It seems that although we are only using
a limited selection of icons from this package
we are seeing the react-icons bundle at 2.61Mb

Switching to use @react-icons/all-files instead
which seems to offer better tree shaking.

The linked GitHub issue thread suggests that
webpack _should_ be able to optimise the bundle
size. However I think owing to our dependency
on an outdated CRA we are not getting those benefits.

https://github.com/react-icons/react-icons/issues/154