fix: use <Button> component for user interaction
We have an existing button component which should
be used to handle user interactions.

The Button element uses semantically correct
HTML elements which improves usability and accessibility.

If the visual styling is a blocker then we can introduce
a new variant, however visual consistency of clickable elements
is a helpful design convention.

Bonus! This removes usage of react-icons within the core
platform and moves knowledge of icons out to the component
library layer.