feat: mocks AuthWrapper
The AuthWrapper component is a useful container component
that given the name of a user checks their roles
and if they match the required role renders the child elements.

This container component does not make sense outside of the
Platform however is used in a lot of components. The container
introduces a dependency on the userStore which is _not_ available
in the Storybook context.

Whilst we could inject a mock store the aim is for components to be pure
functions and have minimal knowledge of external state.

Introducing a mocked version of AuthWrapper unblocks documenting our
components with Storybook that are still with the core platform.